+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Cheapie UI 2024.2.0.1
 
 - I had to fix some stuff.
+- Search still doesn't work
 
 # Cheapie UI 2024.2.0.0